+++ app/static/js/_format-utils.js
@@ -74,3 +74,39 @@ export const formatTime = (seconds) => {
     const m = Math.floor((seconds % 3600) / 60)
     return `${h}:${m.toString().padStart(2, "0")}`
 }
+
+/**
+ * Format degrees to their correct math representation
+ * @param {int} decimalDegree degrees
+ * @returns {string}
+ * @example formatDegrees(21.32123)
+ * // => "21°19′16″"
+ */
+export const formatDegrees = (decimalDegree) => {
+    const degrees = Math.floor(decimalDegree)
+    const minutes = Math.floor((decimalDegree - degrees) * 60)
+    const seconds = Math.round(((decimalDegree - degrees) * 60 - minutes) * 60)
+
+    // Pad single digits with a leading zero
+    const formattedDegrees = degrees < 10 ? `0${degrees}` : `${degrees}`
+    const formattedSeconds = seconds < 10 ? `0${seconds}` : `${seconds}`
+    const formattedMinutes = minutes < 10 ? `0${minutes}` : `${minutes}`
+
+    return `${formattedDegrees}°${formattedMinutes}′${formattedSeconds}″`
+}
+
+/**
+ * Format lat lon in cordinate system. See https://en.wikipedia.org/wiki/Geographic_coordinate_system
+ * @param {L.LatLng} pos position on map
+ * @returns {string}
+ * @example formatLatLon({lat: 21.32123, 35.2134})
+ * // => "21°19′16″N, 35°12′48″E"
+ */
+
+export const formatLatLon = (latLng) => {
+    const lat = formatDegrees(latLng.lat)
+    const lon = formatDegrees(latLng.lng)
+    const latDir = latLng.lat === 0 ? "" : latLng.lat > 0 ? "N" : "S"
+    const lonDir = latLng.lat === 0 ? "" : latLng.lat > 0 ? "E" : "W"
+    return `${lat}${latDir} ${lon}${lonDir}`
+}
