+++ app/static/sass/index/_contextmenu.scss
@@ -50,7 +50,7 @@
     }
 
     bottom: unset !important;
-    left: unset !important;
-    translate: 5px 5px;
+    top: 5px !important;
+    left: 5px !important;
     transition: none;
 }
