+++ app/templates/index/_distance.jinja2
@@ -3,7 +3,7 @@
         <div class="section py-2">
             <div class="row g-1">
                 <div class="col">
-                    <i class="bi bi-signpost-2"></i><span>{{ t('mesaure_distance.distance') }}</span> <span
+                    <i class="bi bi-signpost-2"></i> <span>{{ t('mesaure_distance.distance') }}</span>: <span
                         class="total-distance"></span>
                 </div>
                 <div class="col-auto">
